services/api_bcv/requirements.txt
Django==6.0.4
django-ninja==1.6.2
django-cors-headers==4.9.0
beautifulsoup4==4.14.3
requests==2.33.1
lxml==6.1.0
python-dotenv==1.2.2
dj-database-url==3.1.2
psycopg2-binary==2.9.11
gunicorn==21.2.0
whitenoise==6.12.0
